refactored Preset and Budget goal modals so the title and 'Back and Continue' buttons are their own components
