+++ src/components/BudgetGoalModal.jsx
@@ -1,6 +1,8 @@
 import React, { useState, useEffect} from "react";
+import { Back_Continue } from "./Modal_Components/Back_Continue";
+import { Modal_Title } from "./Modal_Components/Modal_Title";
+
 import { withStyles, makeStyles } from "@material-ui/core/styles";
-import Button from "@material-ui/core/Button";
 import Dialog from "@material-ui/core/Dialog";
 import MuiDialogTitle from "@material-ui/core/DialogTitle";
 import MuiDialogContent from "@material-ui/core/DialogContent";
@@ -141,14 +143,8 @@ export default function BudgetGoal(props) {
         fullWidth={true}
         maxWidth="md"
       >
-        <DialogTitle
-          className="customized-dialog-title"
-          onClose={props.handleClose}
-        >
-          <Typography className="customized-dialog-title">
-            Total goal
-          </Typography>
-        </DialogTitle>
+        <Modal_Title handleClose={props.handleClose} title="Total Goal"/>
+  
         <DialogContent className="content">
           <Typography className="what" variant="h5">
             What is your spending goal?
@@ -197,24 +193,7 @@ export default function BudgetGoal(props) {
             </div>
           </div>
         </DialogContent>
-        <DialogActions className="buttons">
-          <Button
-            className="backBtn"
-            onClick={props.handleClose}
-            variant="outlined"
-            color="primary"
-          >
-            BACK
-          </Button>
-          <Button
-            className="contBtn"
-            onClick={submit}
-            variant="outlined"
-            color="primary"
-          >
-            CONTINUE
-          </Button>
-        </DialogActions>
+        <Back_Continue BackClick={props.handleClose} ContClick={submit}/>
       </Dialog>
     </div>
   );
