+++ src/components/Modal_Components/styles.js
@@ -1,0 +1,12 @@
+export const styles = theme => ({
+    root: {
+      margin: 0,
+      padding: theme.spacing(2)
+    },
+    closeButton: {
+      position: "absolute",
+      right: theme.spacing(1),
+      top: theme.spacing(1),
+      color: "#00000"
+    }
+  });
